refactor(components): split components import path by component domain
