+++ src/components/index.ts
@@ -2,5 +2,5 @@ export * from './common';
 export * from './layouts';
 export * from './modal';
 
-export * from './main';
-export * from './qna';
+//export * from './main';
+//export * from './qna';
